Moved MemberFactory to domain.
